Database Connection
Still working on the database connection. receive the relation "players"
does not exist error when trying to save the player. Worked on trying to
put single quotes around the Players in the select statement but to no
prevail. I tried all the tips people that have the same error have but
not working.